Commit 4
Theme Change drop down menu added and module add drop down menu added.
These buttons will allow the user to change or add to the modules that
are listed in the top 5 for that user.  The theme button will allow the
user to change the color scheme from the standard eastern theme to
something that might better suit their personality.